Get this building with an updated toolchain
This changeset pulls all the logic for Maven publishing out for now.
For the moment folks can clone the repo and build it, or I may publish
releases with a jar and have people install that way.

Otherwise, this is mostly updated the gradle files to match changes to
the DSL and updates to match the latest version of the gradle plugin
API.